Add layer freezing, BioCLIP 2.5 pretrained weights and detector fixes
Added:
  - Add --start-number to python -m birder.tools pack to offset output file numbering
  - Add --freeze-layers to freeze the first N layers using a model's block_group_regex
  - Add --freeze-encoder, --freeze-encoder-stages and --freeze-encoder-layers to train_mim.py
  - Add BioCLIP 2.5 ViT-H/14 pretrained weights (vit_h14_pn_bioclip-v25)

Changed:
  - Test compatibility with PyTorch 2.11

Fixed:
  - Fix mixed torch.Tensor / DTensor _foreach_lerp_ crash in DINOv2 and CAPI FSDP teacher EMA updates
  - Fix Sinkhorn float32 teacher targets colliding with student logits in DINOv2 / Franca mixed-precision training
  - Fix student.backbone attribute access through DDP wrapper in train_dino_v2_dist EMA update
  - Fix FCOS anchor sizes to follow backbone-reported max_stride, including 16- and 32-stride backbones
  - Restore upstream gradient path for RT-DETR v1/v2 deeper auxiliary bbox outputs during training
  - Restore LW-DETR upstream matcher class-cost weight, decoder normalization, IA-BCE clamp, Xavier init and projector downsample path
  - Restore Plain-DETR upstream self-attention mask, stride-aligned upsampling/input projection, top-k inference selection and valid-ratio BoxRPB scaling
  - Update YOLO v4 / v4 Tiny target assignment, ignore-mask computation, box-loss weighting and per-class postprocessing to better match Darknet behavior
  - Fix YOLO v3 Darknet ignore-mask computation, restore YOLO-style box-area weighting, keep per-class detections at postprocess time and clean up _build_targets()

+++ birder/net/ssl/dino_v2.py
@@ -146,8 +146,8 @@ class DINOLoss(nn.Module):
     def forward(
         self, student_output_list: list[torch.Tensor], teacher_out_softmax_centered_list: list[torch.Tensor]
     ) -> torch.Tensor:
-        s = torch.stack(student_output_list, 0)
-        t = torch.stack(teacher_out_softmax_centered_list, 0)
+        s = torch.stack(student_output_list, 0).float()
+        t = torch.stack(teacher_out_softmax_centered_list, 0).float()
         lsm = F.log_softmax(s / self.student_temp, dim=-1)
         loss = -(torch.einsum("tbk,sbk->tsb", t, lsm).mean(-1).sum())
 
@@ -158,9 +158,9 @@ class DINOLoss(nn.Module):
     ) -> torch.Tensor:
         total_loss = 0.0
         for s in student_output_list:
-            lsm = F.log_softmax(s / self.student_temp, dim=-1)
+            lsm = F.log_softmax(s.float() / self.student_temp, dim=-1)
             for t in teacher_out_softmax_centered_list:
-                loss = torch.sum(t * lsm, dim=-1)
+                loss = torch.sum(t.float() * lsm, dim=-1)
                 total_loss -= loss.mean()
 
         return total_loss
@@ -279,8 +279,8 @@ class iBOTPatchLoss(nn.Module):
         masks_weight: torch.Tensor,
         n_masked_patches: Optional[int] = None,
     ) -> torch.Tensor:
-        t = teacher_patch_tokens_masked
-        s = student_patch_tokens_masked
+        t = teacher_patch_tokens_masked.float()
+        s = student_patch_tokens_masked.float()
         loss = torch.sum(t * F.log_softmax(s / self.student_temp, dim=-1), dim=-1)
         if n_masked_patches is not None:
             loss = loss[:n_masked_patches]
@@ -292,7 +292,10 @@ class iBOTPatchLoss(nn.Module):
     def forward_no_masks(
         self, student_patch_tokens: torch.Tensor, teacher_patch_tokens: torch.Tensor, student_masks_flat: torch.Tensor
     ) -> torch.Tensor:
-        loss = torch.sum(teacher_patch_tokens * F.log_softmax(student_patch_tokens / self.student_temp, dim=-1), dim=-1)
+        loss = torch.sum(
+            teacher_patch_tokens.float() * F.log_softmax(student_patch_tokens.float() / self.student_temp, dim=-1),
+            dim=-1,
+        )
         loss = torch.sum(loss * student_masks_flat.float(), dim=-1) / student_masks_flat.sum(dim=-1).clamp(min=1.0)
 
         return -loss.mean()
